Revert "Fix for sound mixer history handling on 64-bit platforms"
This reverts commit 1c7e13e42c0dc3e496a6eb923aaa9d1efeb6cda9.

+++ engine/sound/src/sound.cpp
@@ -1082,7 +1082,7 @@ namespace dmSound
         instance->m_FrameCount = SOUND_MAX_HISTORY + (avail_framecount - used_framecount) + SOUND_MAX_FUTURE;
         assert(instance->m_FrameCount <= SOUND_INSTANCE_STATEFRAMECOUNT);
         uint32_t state_bytes = instance->m_FrameCount * sizeof(float);
-        int32_t state_offset = (int32_t)used_framecount - SOUND_MAX_HISTORY;
+        uint32_t state_offset = used_framecount - SOUND_MAX_HISTORY;
 
         for(uint32_t c=0; c<channels; ++c)
         {
